[merge.c, merge_mono.c] reverted change 1719:1720: Francois will solve himself                         the merge problem on 32-bit machines
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@1763 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/merge.c
@@ -212,9 +212,10 @@ main(int argc, char *argv[])
 #endif
 
     gzip_close(rep.outfile, outname);
-    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%1.0f\n",
+    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%lu\n",
 	    mat.rem_nrows, mat.rem_ncols, mat.rem_nrows-mat.rem_ncols,
-	    mat.weight, (double) mat.rem_nrows * (double) mat.weight);
+	    mat.weight,
+	    ((unsigned long)mat.rem_nrows) * mat.weight);
 #if TEX
     fprintf(stderr, "\\end{verbatim}\n");
 #endif
